+++ src/main/java/com/pi4j/drivers/sensor/environment/scd4x/Scd4xDriver.java
@@ -61,7 +61,7 @@ public class Scd4xDriver implements Sensor {
     }
 
 
-    public void readMeasurement(float[] values) {
+    public void readMeasurement(double[] values) {
         Measurement measurement = readMeasurement();
         values[0] = measurement.temperature;
         values[1] = measurement.humidity;
